only update existing tests
