+++ test/development/acceptance/ReactRefreshLogBox.test.ts
@@ -3,6 +3,7 @@ import { createSandbox } from 'development-sandbox'
 import { FileRef, nextTestSetup } from 'e2e-utils'
 import {
   describeVariants as describe,
+  getStackFramesContent,
   toggleCollapseCallStackFrames,
 } from 'next-test-utils'
 import path from 'path'
@@ -768,5 +769,67 @@ describe.each(['default', 'turbo'])('ReactRefreshLogBox %s', () => {
     )
 
     expect(callStackFrames.length).toBeGreaterThan(9)
+  })
+
+  // TODO: hide the anonymous frames between 2 ignored frames
+  test('should show anonymous frames from stack trace', async () => {
+    await using sandbox = await createSandbox(
+      next,
+      new Map([
+        [
+          'pages/index.js',
+          outdent`
+          export default function Page() {
+            [1, 2, 3].map(() => {
+              throw new Error("anonymous error!");
+            })
+          }`,
+        ],
+      ])
+    )
+
+    const { session, browser } = sandbox
+
+    await session.assertHasRedbox()
+
+    const stack = await getStackFramesContent(browser)
+    expect(stack).toMatchInlineSnapshot(`
+     "at Array.map ()
+     at Page (pages/index.js (2:13))"
+    `)
+  })
+
+  test('should collapse nodejs internal stack frames from stack trace', async () => {
+    await using sandbox = await createSandbox(
+      next,
+      new Map([
+        [
+          'pages/index.js',
+          outdent`
+          export default function Page() {}
+          
+          function createURL() {
+            new URL("/", "invalid")
+          }
+
+          export function getServerSideProps() {
+            createURL()
+            return { props: {} }
+          }`,
+        ],
+      ])
+    )
+
+    const { session, browser } = sandbox
+    await session.assertHasRedbox()
+
+    const stack = await getStackFramesContent(browser)
+    expect(stack).toMatchInlineSnapshot(
+      `"at getServerSideProps (pages/index.js (8:3))"`
+    )
+
+    await toggleCollapseCallStackFrames(browser)
+    const stackCollapsed = await getStackFramesContent(browser)
+    expect(stackCollapsed).toContain('at new URL ()')
   })
 })
